section 5 - identityServer

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -47,12 +48,13 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
     return mapper.Map<AuctionDto>(auction);
   }
 
+  [Authorize]
   [HttpPost]
   public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
   {
     var auction = mapper.Map<Auction>(createAuctionDto);
-    // TODO: Add the current user to the auction as Seller
-    auction.Seller = "test";
+
+    auction.Seller = User.Identity.Name;
 
     context.Auctions.Add(auction);
 
@@ -68,6 +70,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
         new { Id = auction.Id }, newAuction);
   }
 
+  [Authorize]
   [HttpPut("{id}")]
   public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
   {
@@ -76,7 +79,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
 
     if (auction == null) return NotFound();
 
-    // TODO: check seller == username
+    if (auction.Seller != User.Identity?.Name) return Forbid();
 
     auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
     auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -93,6 +96,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
     return BadRequest("Problem saving changes");
   }
 
+  [Authorize]
   [HttpDelete("{id}")]
   public async Task<ActionResult> DeleteAuction(Guid id)
   {
@@ -100,7 +104,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
 
     if (auction == null) return NotFound();
 
-    // TODO: check seller == username
+    if (auction.Seller != User.Identity?.Name) return Forbid();
 
     context.Auctions.Remove(auction);
